Move visualizer types to `metabase-types/api`
