+++ frontend/src/metabase/visualizations/visualizations/DashCardPlaceholder.tsx
@@ -12,8 +12,11 @@ import {
 import { useDispatch } from "metabase/lib/redux";
 import { Button, Flex } from "metabase/ui";
 import { VisualizerModal } from "metabase/visualizer/components/VisualizerModal";
-import type { Dashboard, VirtualDashboardCard } from "metabase-types/api";
-import type { VisualizerVizDefinition } from "metabase-types/store/visualizer";
+import type {
+  Dashboard,
+  VirtualDashboardCard,
+  VisualizerVizDefinition,
+} from "metabase-types/api";
 
 import type { VisualizationProps } from "../types";
