+++ frontend/src/metabase/visualizer/components/DataImporter/DatasetsList/DatasetsListItem.tsx
@@ -10,7 +10,7 @@ import {
   getVisualizerPrimaryColumn,
 } from "metabase/visualizer/selectors";
 import { parseDataSourceId } from "metabase/visualizer/utils";
-import type { VisualizerDataSource } from "metabase-types/store/visualizer";
+import type { VisualizerDataSource } from "metabase-types/api";
 
 import { useVisualizerUi } from "../../VisualizerUiContext";
