+++ frontend/src/metabase/visualizer/visualizations/cartesian.unit.spec.ts
@@ -1,5 +1,6 @@
 import _ from "underscore";
 
+import type { VisualizerVizDefinition } from "metabase-types/api";
 import {
   createMockCategoryColumn,
   createMockColumn,
@@ -7,7 +8,6 @@ import {
   createMockDatetimeColumn,
   createMockNumericColumn,
 } from "metabase-types/api/mocks";
-import type { VisualizerVizDefinition } from "metabase-types/store/visualizer";
 
 import {
   copyColumn,
