+++ frontend/src/metabase/visualizer/visualizations/pie.ts
@@ -14,11 +14,12 @@ import {
   isMetric,
   isNumeric,
 } from "metabase-lib/v1/types/utils/isa";
-import type { Dataset, DatasetColumn } from "metabase-types/api";
 import type {
+  Dataset,
+  DatasetColumn,
   VisualizerDataSource,
   VisualizerVizDefinition,
-} from "metabase-types/store/visualizer";
+} from "metabase-types/api";
 
 import { removeColumnFromStateUnlessUsedElseWhere } from "./utils";
