+++ frontend/src/metabase/visualizer/visualizations/pie.unit.spec.ts
@@ -1,9 +1,9 @@
+import type { VisualizerVizDefinition } from "metabase-types/api";
 import {
   createMockCategoryColumn,
   createMockDataset,
   createMockNumericColumn,
 } from "metabase-types/api/mocks";
-import type { VisualizerVizDefinition } from "metabase-types/store/visualizer";
 
 import { createDataSource } from "../utils";
